LayoutAnimations: Allow user-configurable failure function
Summary:
There's currently a hidden parameter passed to the LayoutAnimation engine for "failure". It's not been used anywhere, ever, since it's a hidden parameter... and it's probably not called from the native side in non-Fabric. But it's been added to the native side of Fabric for feature-completeness and we might as well expose it, the cost seems minimal.

Changelog: [Internal]

Reviewed By: shergin

Differential Revision: D22889455

fbshipit-source-id: 73322776ed21e58a370500ccfde89285f07a9d50